requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
boto3==1.9.82
botocore==1.12.161
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
docutils==0.14
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
more-itertools==7.0.0
numpy==1.16.3
pandas==0.24.2
pluggy==0.11.0
py==1.8.0
pyaml==18.11.0
PyMySQL==0.9.3
pyparsing==2.4.0
pytest==4.0.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
s3transfer==0.1.13
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.1
urllib3==1.24.3
Werkzeug==0.15.4
